@@ -3,30 +3,30 @@
   "version": "0.1.0",
   "private": true,
   "homepage": "http://ShiroKatake.github.io/ytdl-frontend",
+  "proxy": "https://shirokatake-ytdl-backend.herokuapp.com",
   "dependencies": {
     "@types/he": "^1.1.2",
-    "axios": "^0.25.0",
+    "axios": "^0.27.2",
     "bootstrap": "^5.1.3",
-    "react": "^17.0.2",
-    "react-bootstrap": "^2.0.2",
-    "react-dom": "^17.0.2",
+    "react": "^18.1.0",
+    "react-bootstrap": "^2.3.1",
+    "react-dom": "^18.0.0",
     "react-icons": "^4.3.1",
-    "react-scripts": "5.0.0",
-    "typescript": "^4.5.5",
+    "react-scripts": "5.0.1",
+    "typescript": "^4.6.4",
     "web-vitals": "^2.1.4",
-    "ws": "^8.2.3"
+    "ws": "^8.6.0"
   },
   "devDependencies": {
-    "@testing-library/jest-dom": "^5.16.1",
-    "@testing-library/react": "^12.1.2",
-    "@testing-library/user-event": "^13.5.0",
-    "@types/jest": "^27.4.0",
-    "@types/node": "^17.0.11",
-    "@types/react": "^17.0.38",
-    "@types/react-dom": "^17.0.11",
+    "@testing-library/jest-dom": "^5.16.4",
+    "@testing-library/react": "^13.2.0",
+    "@testing-library/user-event": "^14.1.1",
+    "@types/jest": "^27.5.0",
+    "@types/react": "^18.0.9",
+    "@types/node": "^17.0.33",
     "gh-pages": "^3.2.3",
-    "jest": "^27.4.7",
-    "ts-jest": "^27.1.3"
+    "jest": "^27.5.1",
+    "ts-jest": "^27.1.4"
   },
   "scripts": {
     "start": "react-scripts start",
