@@ -2,24 +2,25 @@ import { useEffect, useState } from "react";
 import { ProgressBar } from "react-bootstrap";
 import { useAppContext } from "./context/AppContext";
 import { Button, CurrentVideoInfo, Suggestions, FormatList, Playlist, TextInput } from "./components";
-import { getInfos, getSuggestions, downloadFileFromUrl, getPlaylist } from "./utils/handler/API";
-import { createWebSocketConnection, getYtID, isJson, isUid, isYtList, generateDownloadUrl, generateProgressText, sendMessage } from "./utils";
+import { getInfos, getSuggestions, getPlaylist } from "./utils/handler/API";
+import { createWebSocketConnection, getYtID, tryParseJSON, isYtList, generateProgressText, downloadFromLink } from "./utils";
 import "./App.css";
 
 export const App = () => {
   const { setButtonIsLoading } = useAppContext();
+  console.log(window.location.hostname);
 
   const [inputText, setInputText] = useState<string>("");
-  const [downloadFormat, setDownloadFormat] = useState<string>(localStorage.getItem("format") ? localStorage.getItem("format")! : "mp4");
+  const [downloadFormat, setDownloadFormat] = useState<string>(
+    localStorage.getItem("format") || "mp4"
+  );
 
   const [currentVideoInfo, setCurrentVideoInfo] = useState<any>(null);
   const [suggestions, setSuggestions] = useState<any>([]);
   const [playlistInfo, setPlaylistInfo] = useState<any>([]);
 
-  const [isHidden, setIsHidden] = useState(true);
+  const [progressHidden, hideProgress] = useState(true);
   const [downloadedPercent, setDownloadedPercent] = useState(0);
-  const [downloaded, setDownloaded] = useState(0);
-  const [totalDownloadSize, setTotalDownloadSize] = useState(1);
 
   const [timeoutFunctionId, setTimeoutFunctionId] = useState<NodeJS.Timeout>();
 
@@ -45,81 +46,94 @@ export const App = () => {
     if (ytId) {
       await download(ytId);
     } else if (isYtList(inputText)) {
-      setIsHidden(true);
+      hideProgress(true);
       await getPlaylist(inputText, setPlaylistActive);
+      setButtonIsLoading(false);
     } else {
-      setIsHidden(true);
+      hideProgress(true);
       await getSuggestions(inputText, setSuggestionsActive);
+      setButtonIsLoading(false);
     }
-    setButtonIsLoading(false);
   };
 
   const download = async (videoId: string) => {
-    setDownloadedPercent(0);
-    if (timeoutFunctionId) {
-      clearInterval(timeoutFunctionId);
-    }
-    setButtonIsLoading(true);
-    setIsHidden(false);
-
     try {
       const videoUrl = videoId || inputText;
       if (!videoUrl) {
-        throw "Invalid video";
+        throw new Error("Invalid video.");
       }
 
+      setDownloadedPercent(0);
+      if (timeoutFunctionId) {
+        clearInterval(timeoutFunctionId);
+      }
+      setButtonIsLoading(true);
+      hideProgress(false);
+
       const data = await getInfos(videoUrl);
-
-      const downloadUrl = generateDownloadUrl(videoUrl, downloadFormat);
-      const socket = createWebSocketConnection();
-      let uid = "";
-
-      socket.addEventListener("message", (event) => {
-        uid = isUid(event.data);
-        if (isJson(event.data)) {
-          const downloadProgress = JSON.parse(event.data);
-          //The encoding process is 75% of the download process
-          // The other 25% will be sending to the front-end for download
-          setDownloadedPercent((downloadProgress.downloaded / downloadProgress.total) * 75);
-          setDownloaded(downloadProgress.downloaded);
-          setTotalDownloadSize(downloadProgress.total);
-        }
-      });
-      await sendMessage(socket, uid);
-
       setCurrentVideoInfo(data.videoDetails);
 
       console.log("Starting download . . .");
-      const filename = `${data.videoDetails.title}.${downloadFormat}`;
-      await downloadFileFromUrl(downloadUrl!, uid, setDownloadedPercent, filename);
-    } catch (error) {
-      console.error(error);
-    }
 
-    setButtonIsLoading(false);
-    setTimeoutFunctionId(
-      setTimeout(() => {
-        setIsHidden(true);
-        setDownloadedPercent(0);
-      }, 3000)
-    );
+      const socket = createWebSocketConnection();
+
+      socket.addEventListener("open", () => {
+        const downloadData = {
+          v: videoUrl,
+          format: downloadFormat,
+        }
+        socket.send(JSON.stringify(downloadData));
+      });
+
+      socket.addEventListener("message", (event) => {
+        const downloadProgress = tryParseJSON(event.data);
+        if (downloadProgress) {
+          setDownloadedPercent(
+            Math.round((downloadProgress.downloaded / downloadProgress.total) * 100)
+          );
+        } else {
+          const url = window.URL.createObjectURL(new Blob([event.data]));
+          const filename = `${data.videoDetails.title}.${downloadFormat}`;
+          socket.close();
+
+          downloadFromLink(url, filename);
+
+          setButtonIsLoading(false);
+          setTimeoutFunctionId(
+            setTimeout(() => {
+              hideProgress(true);
+              setDownloadedPercent(0);
+            }, 3000)
+          );
+        }
+      });
+    } catch (error: any) {
+      console.error(error.message);
+      setButtonIsLoading(false);
+      setTimeoutFunctionId(
+        setTimeout(() => {
+          hideProgress(true);
+          setDownloadedPercent(0);
+        }, 5000)
+      );
+    }
   };
   // prettier-ignore
   return (
     <>
       <section className="search-section">
-        <TextInput inputText={inputText} setInputText={setInputText} />
+        <TextInput inputText={inputText} setInputText={setInputText} onEnter={checkInputText} />
         <ProgressBar
-          hidden={isHidden}
-          striped variant="success"
+          animated={downloadedPercent === 100 ? true : false}
+          hidden={progressHidden}
+          striped
+          variant="success"
           now={downloadedPercent}
-          label={generateProgressText(downloadedPercent, downloaded, totalDownloadSize)}
+          label={generateProgressText(downloadedPercent)}
           style={{ width: "100%", height: "30px", lineHeight: "30px" }}
         />
         <FormatList downloadFormat={downloadFormat} setDownloadFormat={setDownloadFormat} />
-        <Button main onClick={checkInputText}>
-          Search
-        </Button>
+        <Button main onClick={checkInputText}>Search</Button>
       </section>
       <section className="downloading-section">
         {currentVideoInfo && <CurrentVideoInfo currentVideoInfo={currentVideoInfo} />}
