@@ -1,20 +1,18 @@
-import { act, cleanup, fireEvent, render, screen } from '@testing-library/react';
-import { Playlist } from './Playlist';
-import mockedPlaylist from '../../mocks/mockedPlaylist';
+import { act, cleanup, fireEvent, render, screen } from "@testing-library/react";
+import { Playlist } from "./Playlist";
+import mockedPlaylist from "../../__mocks__/mockedPlaylist";
 
 const mockedDownload = jest.fn((videoId: string) => videoId);
 
-describe('Playlist', () => {
+describe("Playlist", () => {
   beforeEach(() => {
-    render(
-      <Playlist playlistInfo={mockedPlaylist} download={mockedDownload}/>
-    );
+    render(<Playlist playlistInfo={mockedPlaylist} download={mockedDownload} />);
     mockedDownload.mockClear();
   });
 
   afterEach(cleanup);
 
-  it('should pass the correct video id on click download', () => {
+  it("should pass the correct video id on click download", () => {
     const downloadButton = screen.getByTestId("downloadButton-0");
     fireEvent.click(downloadButton);
 
@@ -22,7 +20,8 @@ describe('Playlist', () => {
     expect(mockedDownload).toHaveBeenCalledWith("videoId0");
   });
 
-  it('should pass the correct video id list on click download-selected', () => {
+  // prettier-ignore
+  it("should pass the correct video id list on click download-selected", () => {
     const checkboxAll = screen.getByTestId("checkbox-universal");
     const checkboxSingle = screen.getByTestId("checkbox-0");
     const checkboxSingleOther = screen.getByTestId("checkbox-1");
@@ -32,18 +31,18 @@ describe('Playlist', () => {
     fireEvent.click(downloadAllButton);
 
     expect(mockedDownload).toHaveBeenCalledTimes(1);
-    expect(mockedDownload).toHaveBeenCalledWith("{\"0\":false,\"1\":true}");
+    expect(mockedDownload).toHaveBeenCalledWith('{"0":false,"1":true}');
 
     act(() => { fireEvent.click(checkboxSingle) });
     fireEvent.click(downloadAllButton);
 
     expect(mockedDownload).toHaveBeenCalledTimes(2);
-    expect(mockedDownload).toHaveBeenCalledWith("{\"0\":true,\"1\":true}");
+    expect(mockedDownload).toHaveBeenCalledWith('{"0":true,"1":true}');
 
     act(() => { fireEvent.click(checkboxAll) });
     fireEvent.click(downloadAllButton);
 
     expect(mockedDownload).toHaveBeenCalledTimes(3);
-    expect(mockedDownload).toHaveBeenCalledWith("{\"0\":false,\"1\":false}");
+    expect(mockedDownload).toHaveBeenCalledWith('{"0":false,"1":false}');
   });
 });
