@@ -4,7 +4,11 @@ import { IAppContext } from "./IAppContext";
 export const AppContext = createContext<IAppContext>({} as IAppContext);
 export const useAppContext = () => useContext(AppContext);
 
-export const AppContextProvider: React.FC = ({ children }) => {
+interface AppContextProviderProps {
+  children?: React.ReactNode;
+}
+
+export const AppContextProvider: React.FC<AppContextProviderProps> = ({ children }) => {
   const [inputText, setInputText] = useState("");
   const [isLoading, setIsLoading] = useState(false);
 
