@@ -3,7 +3,7 @@ import { hostname } from "../handler/hostname";
 
 export const generateDownloadUrl = (videoId: string, format = "mp4") => {
   if (!downloadableFormats.includes(format)) {
-    throw "Invalid url.";
+    throw new Error("Invalid format.");
   }
   return `${hostname}/download?v=${videoId}&format=${format}`;
 };
