@@ -1,12 +1,7 @@
-export const generateProgressText = (downloadProgress: number, downloaded: number, totalSize: number) => {
-  let text = `Fetching . . . ${toMB(downloaded)}MB / `;
-  if (downloadProgress > 75) {
-    text = "Preparing download . . . ";
+export const generateProgressText = (downloadedPercent: number) => {
+  let text = `Fetching . . . ${downloadedPercent}%`;
+  if (downloadedPercent === 100) {
+    text = "Done! Preparing download . . .";
   }
-  if (downloadProgress === 100) {
-    text = "Ready! ";
-  }
-  return text + `${toMB(totalSize)}MB`;
+  return text;
 };
-
-const toMB = (val: number) => (val / 1024 / 1024).toFixed(2);
